refactor: category layout structure
